feed handler: ITCH message handlers, length validation, end-to-end fixture test
- on_add/delete/execute/cancel/replace parse A/F/D/E/C/X/U into the book
- order_ref_t widened to uint64_t (ITCH refs exceed 32 bits intraday)
- per-message length table + size asserts guard malformed frames
- messages_ counts book-affecting messages, exposed via messages()
- test replays real ITCH fixture: 293482 frames, 602 SPY msgs, uncrossed book

+++ include/hft/types.hpp
@@ -31,7 +31,7 @@ using qty_t = std::int64_t;       // quantity in smallest tradeable unit
 using order_id_t = std::uint64_t;
 using SymbolId = std::uint32_t;
 enum class Side : std::uint8_t { Buy, Sell };
-using order_ref_t = std::uint32_t;
+using order_ref_t = std::uint64_t;
 
 struct StringHash {
     using is_transparent = void;
